+++ src/main/java/net/succ/solar_punk/ModTags.java
@@ -11,10 +11,5 @@ public class ModTags {
                 Registries.BIOME,
                 ResourceLocation.fromNamespaceAndPath(SolarPunk.MODID, "has_salt_deposits")
         );
-
-        public static final TagKey<Biome> HAS_GEYSERS = TagKey.create(
-                Registries.BIOME,
-                ResourceLocation.fromNamespaceAndPath(SolarPunk.MODID, "has_geysers")
-        );
     }
 }