+++ src/main/java/net/succ/solar_punk/worldgen/ModBiomeModifierTypes.java
@@ -1,0 +1,21 @@
+package net.succ.solar_punk.worldgen;
+
+import com.mojang.serialization.MapCodec;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.neoforge.common.world.BiomeModifier;
+import net.neoforged.neoforge.registries.DeferredHolder;
+import net.neoforged.neoforge.registries.DeferredRegister;
+import net.neoforged.neoforge.registries.NeoForgeRegistries;
+import net.succ.solar_punk.SolarPunk;
+
+public class ModBiomeModifierTypes {
+    public static final DeferredRegister<MapCodec<? extends BiomeModifier>> BIOME_MODIFIER_SERIALIZERS =
+            DeferredRegister.create(NeoForgeRegistries.Keys.BIOME_MODIFIER_SERIALIZERS, SolarPunk.MODID);
+
+    public static final DeferredHolder<MapCodec<? extends BiomeModifier>, MapCodec<ConfigurableGeyserBiomeModifier>> CONFIGURABLE_GEYSER_BIOMES =
+            BIOME_MODIFIER_SERIALIZERS.register("configurable_geyser_biomes", () -> ConfigurableGeyserBiomeModifier.CODEC);
+
+    public static void register(IEventBus eventBus) {
+        BIOME_MODIFIER_SERIALIZERS.register(eventBus);
+    }
+}